Custom method to get the number of columns in a row.
We use this new method in ExcelReader.tableGetNextRow instead of apache
poi Row.getLastCellNum. If there was any formatting, etc in a cell in
the given row, this would be included in the getLastCellNum. In some
cases this was returnig upward of 30,000 cells. When creating the sqlite
db, this was causing error as we create 1 table column per spreadsheet
column.