Create Ezid for most recent Bcid.
When a Bcid was being created, the corresponding ezid wasn't created at
the same time. The ezid would be created upon the next Bcid creation.
This had to do with the createBcid method being transactional. The db
wasn't flushed before the createBcidsEZIDs method being called. Thus the
reason the ezid is created for the previous bcid.