Store numRows in the ExcelReader
We we previously counting the rows via apache poi everytime
ExcelReader.getNumRows() was called. Now we only count with poi if we
haven't previously counted the rows for the active sheet. Since
getNumRow() was being called in for loops, this improved performance of
the loop 1100x with a dataset with 1000 rows. Since during uploading,
the could be multiple for loops calling getNumRows(), I expect this to
dramatically speed up validation/uploading.