Large refactoring of validation and uploading code
Created a FileManager interface and implementations. This includes a
special implementation DatasetFileManager, that lives in
biocode-fims-commons. This refactoring makes the validation and
uploading code use "plugins". We can now create a new FileManager
implementation for each filetype that we support.

The default file types are FIMS datasets. Additional FileManagers all
work with/attach data to the Dataset. Some examples are sequence data
(FASTA and FASTQ files) as well as photos.

Each FileManager can also specify a persistenceManager. This allows us
to have more flexibility with client FIMS instances in that we can
support the clients persistence backend of choice.

Removed fasta related classes to seperate packages.