Implement api varsioning
This commit introduces a pipeline for transforming REST requests and
responses up and down the api version ladder. This allows us to
update the api without breaking our existing api.

There is a filter that pulls the requested api version from the url and
places it in a header. Thus, the api version can be specified with the
header "Api-Version" or via the url. The default api version is 1.0 to
remain backwords compatible