Created File Writers to write json objects to files.
This is to replace the current QueryWriter class. JsonWriters are more
versitle. They accept a ArrayNode of json resources, and a List of
JsonFieldTransform objects. The JsonFieldTransform objects contain
information about how each field in the json resource should be written.
This can include flatening the json resource, as well as renaming the
field into a more human readable name