Added scripts to fix bcid timestamps if they get autoupdated on accident
When we select the latest dataset, we use the ts to fetch the most
recent. Bcids have auto update on the ts column. If we ever update a
group of bcids and forget to turn off the ts auto_update, this will
break fims as we won't be able to fetch the lastest dataset.

These scripts will update the ts column from the ezid entry if possible.
If not, then we will create an arbitrary ts with a later date as the
bcidId increases;