Moved ElasticSearch Config from FimsAppConfig
Seperated out the elasticsearch specific beans into its own class. This
allows us to import the ElasticSearchAppConfig.class into *-fims
projects, and not have to specify the beans in every project, while
keeping the elastic search beans out of the projects that don't use
elasticsearch.

Made ProjectRestService an abstract class, and removed esClient. This
allows us to extend and override the class in the *-fims instances,
keeping elasticsearch out of the core rest classes