Fix bug with subResource not being spring beans
We need to explicitly register subresource with jersey in order to get
the resource with the SpringComponentProvider. Otherwise, the
subresources are hk2 beans. This is a problem because the
VersionTransformer advice only works with spring beans.

Also added a check upon jersey application init to verify that all REST
services/resources are spring beans.