Replaced OAuthTokenReposityImpl.getUser native query with jpa criteria query
This allows us to fetch the user with a jpa entityGraph to dynamically
load the user entity lazy collections, when accessing a REST service
using oAuth.

Added a @UserEntityGraph annotation. This is used to specify a
entityGraph to load when fetching the authenticated User. This
is useful in prventing multiple queries for User when we need to work
with lazily loaded collections.

Added authorization security to ExpeditionResource.listExpeditions.