Modified listProjects rest resource
Created a ProjectResource to list all the projects for the request user.
This includes public projects for the current appRoot url and any
private projects that the user is a member of for the current appRoot.

Added a method to UserProjectResource to get all of the projects a user
is a member of at the current projectUrl.

implemented a version transformer to maintain backwards compatibility.