Beginnings of a more generic DataReader interface
Moving away from the TabularDataReader to a more generic DataReader
interface. Implementing classes are responsible for converting a file
into a list of RecordSets, based on the Project Mapping information.

Created a RecordSet to store all Records for a particular Entity in a
Mapping. This is necessary to support multi-entity spreadsheets. This
will also provide the basis for validation. Currently we validate a
"spreadsheet" at a time, however, we need to move to entity based
validation.

Created a Record interface and GenericRecord implementation. This will
provide a more OO way to work with data throughout the application.