Merge Bcid branch.
Resolver code is currently commented out until we figure out how best to
handle that
started transitioning away from SettingManager to a FimsProperties class. This commit is incomplete and contains build errors

Start removing bcid/ezid code

This involves creating a BcidTmp class to house the current Bcid
information. This will be duplicated data from the Biocode-BCID
repository. Biocode-BCID will be the source db for Bcids and will be
creating ezids.

The BcidTmp class will house information that we currently need, but is
removed in the postgresql release. This includes subResourceTypes,
sourceFiles, and graphs. This is needed to run the current system, but
will not be needed in the near future.

The BcidServce will first mint a Bcid with the Biocode-BCID repository
and will then persit the BcidTmp into the fims instance db. This
duplication will be removed in the postgres release.

After successful migration of the Bcid system for all fims instances,
we will begin the larger postgres migration across the different
instances.

started implementing a BcidRepository

add BcidRepository bean

add publisher to Bcid and a few mug fixes

make FimsProperties extendable

getting things sorted out after the cherry-picks