Don't create expedition on upload
Removing the ability to create an expedition when uploading data. This
is a mixing of responsiblities and should be a seperate api call.

Disabled the null expeditionCode check, to allow validating data before
creating an expedition. This needs to be tested more, specifically the
case where there are parent entities invloved.